WIP reorganized dice types in YahtzeeKit.
