+++ YahtzeeKit/Sources/YahtzeeKit/Dice/DiceCup.swift
@@ -1,160 +1,9 @@
 //
-//  Dice.swift
+//  DiceCup.swift
+//  YahtzeeKit
 //
+//  Created by Tom Hartnett on 3/20/25.
 //
-//  Created by Tom Hartnett on 6/20/24.
-//
-
-import Foundation
-
-public enum DieSlot {
-    case one
-    case two
-    case three
-    case four
-    case five
-}
-
-public enum DieValue: Int, CaseIterable {
-    case one = 1
-    case two = 2
-    case three = 3
-    case four = 4
-    case five = 5
-    case six = 6
-
-    public init(integer: Int) {
-        self = .init(rawValue: integer) ?? .one
-    }
-
-    public static func random() -> DieValue {
-        let random = Int.random(in: 1...6)
-        return DieValue(integer: random)
-    }
-}
-
-public struct Die {
-    public var value: DieValue?
-    public var isHeld: Bool = false
-
-    public init(value: DieValue? = nil, isHeld: Bool = false) {
-        self.value = value
-        self.isHeld = isHeld
-    }
-
-    mutating func roll() {
-        guard !isHeld else { return }
-
-        value = DieValue.random()
-    }
-
-    mutating func hold() {
-        guard value != nil else { return }
-
-        isHeld.toggle()
-    }
-}
-
-public struct DiceValues: Equatable {
-    public let value1: DieValue
-    public let value2: DieValue
-    public let value3: DieValue
-    public let value4: DieValue
-    public let value5: DieValue
-
-    public var isThreeOfAKind: Bool {
-        // 1 1 1 3 4
-        let set: Set<Int> = [
-            value1.rawValue,
-            value2.rawValue,
-            value3.rawValue,
-            value4.rawValue,
-            value5.rawValue
-        ]
-        return set.count <= 3
-    }
-
-    public var isFourOfAKind: Bool {
-        // 1 1 1 1 4
-        let set: Set<Int> = [
-            value1.rawValue,
-            value2.rawValue,
-            value3.rawValue,
-            value4.rawValue,
-            value5.rawValue
-        ]
-        return set.count <= 2
-    }
-
-    public var isFullHouse: Bool {
-        // 1 1 1 4 4
-        let set: Set<Int> = [
-            value1.rawValue,
-            value2.rawValue,
-            value3.rawValue,
-            value4.rawValue,
-            value5.rawValue
-        ]
-        return set.count == 2
-    }
-
-    public var isSmallStraight: Bool {
-        // 1 2 3 4 4
-        let set: Set<Int> = [
-            value1.rawValue,
-            value2.rawValue,
-            value3.rawValue,
-            value4.rawValue,
-            value5.rawValue
-        ]
-        return set.count >= 4
-    }
-
-    public var isLargeStraight: Bool {
-        // 1 2 3 4 5
-        let set: Set<Int> = [
-            value1.rawValue,
-            value2.rawValue,
-            value3.rawValue,
-            value4.rawValue,
-            value5.rawValue
-        ]
-        return set.count == 5
-    }
-
-    public var isYahtzee: Bool {
-        // 1 1 1 1 1
-        [value1, value2, value3, value4, value5].allSatisfy({ $0 == value1 })
-    }
-
-    init(
-        _ value1: DieValue,
-        _ value2: DieValue,
-        _ value3: DieValue,
-        _ value4: DieValue,
-        _ value5: DieValue
-    ) {
-        self.value1 = value1
-        self.value2 = value2
-        self.value3 = value3
-        self.value4 = value4
-        self.value5 = value5
-    }
-
-    init(
-        _ value1: DieValue?,
-        _ value2: DieValue?,
-        _ value3: DieValue?,
-        _ value4: DieValue?,
-        _ value5: DieValue?
-    ) {
-        self.value1 = value1 ?? .one
-        self.value2 = value2 ?? .one
-        self.value3 = value3 ?? .one
-        self.value4 = value4 ?? .one
-        self.value5 = value5 ?? .one
-    }
-}
 
 public struct DiceCup {
     public var values: DiceValues? {
@@ -594,18 +443,5 @@ public struct DiceCup {
         die3.isHeld = false
         die4.isHeld = false
         die5.isHeld = false
-    }
-}
-
-extension Array {
-    subscript(safe index: Int) -> Element? {
-        indices.contains(index) ? self[index] : nil
-    }
-}
-
-extension Array {
-    func randomElement(where predicate: (Element) -> Bool) -> Element? {
-        let filteredArray = filter(predicate)
-        return filteredArray.isEmpty ? nil : filteredArray.randomElement()
     }
 }
