+++ YahtzeeKit/Sources/YahtzeeKit/Dice/DieSlot.swift
@@ -1,0 +1,14 @@
+//
+//  DieSlot.swift
+//  YahtzeeKit
+//
+//  Created by Tom Hartnett on 3/20/25.
+//
+
+public enum DieSlot {
+    case one
+    case two
+    case three
+    case four
+    case five
+}
